Further split of client and service with some files we hadn't decided on previously.
