+++ src/org/mozilla/mozstumbler/client/mapview/MapActivity.java
@@ -25,8 +25,7 @@ import java.util.List;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.mozilla.mozstumbler.client.mapview.Searcher;
-import org.mozilla.mozstumbler.blocklist.BSSIDBlockList;
+import org.mozilla.mozstumbler.service.blocklist.BSSIDBlockList;
 import org.mozilla.mozstumbler.BuildConfig;
 import org.mozilla.mozstumbler.R;
 import org.mozilla.mozstumbler.client.MainActivity;
