+++ src/org/mozilla/mozstumbler/service/datahandling/StumblerBundleReceiver.java
@@ -1,4 +1,4 @@
-package org.mozilla.mozstumbler.service;
+package org.mozilla.mozstumbler.service.datahandling;
 
 import android.content.BroadcastReceiver;
 import android.content.ContentValues;
@@ -8,8 +8,6 @@ import android.util.Log;
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
-import org.mozilla.mozstumbler.service.StumblerBundle;
-import org.mozilla.mozstumbler.DatabaseContract;
 
 public final class StumblerBundleReceiver extends BroadcastReceiver {
     private static final String LOGTAG = StumblerBundleReceiver.class.getName();
