+++ src/org/mozilla/mozstumbler/service/utils/NetworkUtils.java
@@ -1,4 +1,4 @@
-package org.mozilla.mozstumbler;
+package org.mozilla.mozstumbler.service.utils;
 
 import android.content.Context;
 import android.net.ConnectivityManager;
